+++ ng_trajectory/penalizers/none/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3.6
+"""Dummy penalizer.
+
+This effectively disables the penalizer, as it passes any candidate.
+"""
+from .main import init, penalize, INVALID_POINTS  # noqa: F401