+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Criterions
     - [**NEW**] _Manual_
         - Criterion 'manual' requests user to specify fitness value for each candidate.
+- Penalizers
+    - [**NEW**] _None_
+        - Dummy penalizer that allows any candidate.
 - Segmentators
     - Function `filterPoints` to filter points outside of the map.
 - 'ng_generate_data': Save image of generated friction map when verbosity is higher.